+++ pkg/batch/scan.go
