+++ pkg/batch/scan_internal_test.go
